// dmgcc/02_WebFrontEnd/src/saga/Types/Finance.Types.ts
export interface FinanceState {
    loading: boolean,
    items: {},
    errors:{},
    getCostCentreData?:any,
    resCostCentreData?:any,
    downloadCentreData?:any,
    getIOMappingData?:any,
    getForoxRateData?:any,
    getRateCardData?:any,
    getvendorData?:any,
    getIODumpData?:any,
    getRateCardMasterData?:any,
    resFileObject?:any,
    resMarketData?:any,
    putResMarketData?:any
    resIoCcChartData?:any,
    resCostCenterBlock?:any,
    resIONumber?:any
}